included sndfile lib, aparently i needed a seperate commit for every lib file and header seperately

+++ Engine/AudioManager.cpp
@@ -80,7 +80,7 @@ int Sound::SoundPaStreamCallback(
 		}
 		scale*=self->volume/100.0;
 		for(sf_count_t m=0;m<static_cast<unsigned int>(readCount/channels/self->pitch);m++) {
-			static_cast<float *>(outputBuffer)[m]=data[((unsigned int)(m*self->pitch))*channels]*scale;
+			static_cast<float *>(outputBuffer)[m]=static_cast<float>(data[((unsigned int)(m*self->pitch))*channels]*scale);
 		}
 	} else { // If no more data, stop the stream
 		if(self->loop) {
